# Conflicts:
#	README-en_US.md
#	README-zh_TW.md
#	README.md
#	app/src/main/java/com/imcys/bilibilias/home/ui/activity/HomeActivity.kt


Former-commit-id: 8f75b34e9f1e09869aaaa44baf65bea1b8391d92

+++ common/src/main/java/com/imcys/bilibilias/common/base/model/common/BangumiFollowList.kt
@@ -30,7 +30,7 @@ data class BangumiFollowList(
             val season_title: String = "",
             val evaluate: String = "",
             val subtitle: String = "",
-            val first_ep: Int = 0,
+            val first_ep: Long = 0,
             val mode: Int = 0,
             val url: String = "",
             val badge_info: BadgeInfoBean = BadgeInfoBean(),
